webserver/requirements_webserver.txt
# Web Framework and ASGI Server
fastapi>=0.109.0
uvicorn>=0.27.0

# Database Connection
psycopg2-binary>=2.9.9

# Deep Learning Framework
torch>=2.2.0
torchvision>=0.17.0

# Data Processing
numpy>=1.24.0
Pillow>=10.0.0